requirements.txt
beautifulsoup4
faiss-cpu
google-api-python-client
google-auth
google-auth-oauthlib
groq
numpy
openai
python-dotenv
sqlalchemy
